Fixed speed data types. Final Code.

+++ self-driving-car.ino
@@ -88,11 +88,11 @@ void setup() {
 
 // Necessary variables for car operation.
 unsigned int thresholdDistance = 0;
-uint8_t rotationSpeed = 0;
-uint8_t rotationOffset = 0;
-uint8_t rightSpeed = 0;
-uint8_t leftSpeed = 0;
-uint8_t mode = 0;
+uint16_t rotationSpeed = 0;
+uint16_t rotationOffset = 0;
+uint16_t rightSpeed = 0;
+uint16_t leftSpeed = 0;
+uint16_t mode = 0;
 unsigned long remoteCode = 0;
 void loop() {
   switch(remoteCode)
